requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.3.6
Pillow==9.0.1
psycopg2-binary==2.9.3
pysqlite3==0.4.7
pytz==2022.1
sqlparse==0.4.2
